plants/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.6.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
timeago==1.0.16
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
aiohttp==3.9.3
scikit-learn==1.4.1.post1
numpy==1.26.4
schedule==1.2.1
requests==2.31.0
